Commonizing ConvertMetdataReference
