+++ src/Microsoft.Dnx.Compilation.CSharp.Common/project.json
@@ -10,10 +10,15 @@
     "frameworks": {
         "dnx451": {
             "frameworkAssemblies": {
+                "System.IO": "4.0.0.0",
                 "System.Runtime": ""
             }
         },
-        "dnxcore50": { }
+        "dnxcore50": {
+            "dependencies": {
+                "System.IO": "4.0.11-*"
+            }
+        }
     },
 
     "scripts": {
